% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptHeader.R
\name{ScriptHeader}
\alias{ScriptHeader}
\title{Header Template for Script Files}
\usage{
ScriptHeader()
}
\value{
Template Header for Scripts
}
\description{
Quick printout of an XML style header for script files.
This information provided in the header file will be used to populate the
data returned to the user when they request a script file list.
}
\examples{

#ScriptHeader()
}
