% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowsSetup.R
\name{SetupWindowsTrigger}
\alias{SetupWindowsTrigger}
\title{Setup MobileTrigger Folder Structure and Files}
\usage{
SetupWindowsTrigger(path, SMTP.Settings, Mail.To, Mail.From)
}
\arguments{
\item{path}{string, top level folder called the [TriggerPath]}

\item{Mail.To}{string, e-mail address you want MobileTriggers to respond to}

\item{Mail.From}{string, e-mail address you want MobileTriggers to use to send.}

\item{SMTP.Server}{string, SMTP server for your e-mail account}

\item{SMTP.Port}{integer, port number}

\item{SMTP.User}{string, user name}

\item{SMTP.Password}{string, password}
}
\value{
Sets up a folder and file structure at the [TriggerPath].
}
\description{
The heart of the MobileTrigger package. You specify where you
want your tigger folder and some e-mail information and this function.
This function builds all the scripts you need to get started. At this time,
the function works for windows users.
}
\examples{

#################################
# Setting of MobileTriggers     #
#################################
# SetupWindowsTrigger(path="c:/triggers",
#                     Mail.To = "Your.Email@mobile.com",
#                     Mail.From = "R.Triggers@desktop.com",
#                     SMTP.Settings=list(
#                         host.name = 'smtp.office365.com',
#                         port = 587,
#                         user.name = 'someuser@outlook.com',
#                         passwd = 'password', tls = TRUE)
# )
}
