% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestTriggers.R
\name{testTriggers}
\alias{testTriggers}
\title{Test Mobile Triggers}
\usage{
testTriggers(TestWhat = "Lists", path = NULL, Mail.From = NULL,
  Mail.To = NULL)
}
\arguments{
\item{TestWhat}{string, options: "Lists", "RunModels", "RunScripts", "RunReports"}

\item{path}{string, top level folder called the [TriggerPath]}

\item{Mail.From}{string, e-mail address your mail client is set to respond to.}

\item{Mail.To}{string, e-mail address of machine running R and mobileTriggerts}
}
\value{
Sends test messages to your email client.
}
\description{
Tests to make sure communication between R and
your e-mail client is working as expected with MobileTriggers.
}
\examples{

#################################
# Test Triggers                 #
#################################
}
