% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerMSG.R
\name{TriggerMSG}
\alias{TriggerMSG}
\title{Send message from mobile triggers}
\usage{
TriggerMSG(path, subject = NULL, body = NULL, html = F,
  authenticate = T, debug = F, ...)
}
\arguments{
\item{path}{string, top level folder called the [TriggerPath]}

\item{subject}{string, subject text of the message}

\item{body}{string, body text of the message}

\item{html}{boolean, use HTML or plain text.}

\item{authenticate}{boolean, use authentication for SMTP signin}
}
\value{
message to be sent to mobile email client.
}
\description{
Function is a wrapper for mailR::send.mail() function.
}
\examples{

####################################################
# Setting of MobileTriggers With ThunderBird Rules #
####################################################
# the function internals will call your mailsettings
# from the mailsettings.R file in your trigger root folder

# TriggerMSG(path = 'c:/triggers',
#           body = "hello world",
#           subject = "Script List",
#           html = F, authenticate = T
# )
}
